Added User Auth Checks for Deleting Blogs and Modified Tests for it

+++ controllers/blogs.js
@@ -55,6 +55,23 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
 	const blogId = request.params.id;
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: 'token invalid' });
+	}
+
+	const user = await User.findById(decodedToken.id);
+	const blog = await Blog.findById(blogId);
+
+	if (!blog) {
+		response.status(204).end();
+	}
+
+	if (blog.user.toString() !== user.id) {
+		return response.status(401).json({ error: 'Unauthorized Access' });
+	}
+
 	await Blog.findByIdAndDelete(blogId);
 	response.status(204).end();
 });
